Split Application State into partial classes.
